Add support for the renamed Deelocity Maven Extension
The Gradle Enterprise Maven Extension was renamed to the Develocity Maven
Extension in version 1.21.  The API was also updated for hte product name
change. Unfortunately, this broke compatibility with the Build Validation
Scripts. This commit fixes the incompatibility.

The fix works by updating the configure-gradle-enterprise-maven-extension to
define and register two listeners: a GradleEnterpriseListener and a
DevelocityListener.  If an older version of the Gradle Enterprise Maven
Extension is used, it will load the GradleEnterpriseListener.  If a newer
version of the Develocity Maven Extension is used, then it will load the
DevelocityListener. The Develocity Maven Extension Adapters are used to
eliminate duplication between the two different listeners.

The develocity-maven-extention-adapters jar has to be added to the Maven
extension classpath.  The Bash scripts were updated to do so, and the Gradle
build scripts were updated to inject the adaptor version into the Bash scripts.

ALternatively, we could:
  1. Create a fat jar containing the listeners and the adapters

  2. Update the bash scripts to search the maven lib dir for the jar, and
     use the latest one it finds.
