+++ components/configure-gradle-enterprise-maven-extension/build.gradle.kts
@@ -9,7 +9,8 @@ repositories {
 dependencies {
     compileOnly("org.apache.maven:maven-core:3.9.6")
     compileOnly("org.codehaus.plexus:plexus-component-annotations:2.2.0")
-    compileOnly("com.gradle:gradle-enterprise-maven-extension:1.18.4")
+    compileOnly("com.gradle:develocity-maven-extension:1.21.3")
+    implementation("com.gradle:develocity-maven-extension-adapters:${rootProject.extra["develocityMavenExtensionAdaptersVersion"]}")
 }
 
 description = "Maven extension to capture the build scan URL"
